Day 10, part 1
I've got a bit to do today so I wasn't able to finish part 2, but I wouldn't be surprised if people actually didn't finish this part, since figuring out polygon overlaps are a bit crazy. It took me a little bit of "being dumb" to realize that there were particularly specific points of the grid that had to be connected, not just all directions.

At the end of the day though, I will probably eventually do the second part, but it may take some time and a long think to figure it out for real.