Day 11
Technically what I did is just BFS, but, in a way, it is faster, right?

This wasn't that bad, mostly because I did the cool thing and optimized the emptyRows and emptyCols first before doing this, so Part 2 was an absolute breeze by just changing one number. Honestly, if every day felt like this, I'd be somewhat happy, because I just hated the loop question from the other day.