Day 12
I copied a lot of this today, but I tried to do part 2 on my own -- but realizing the calls were taking too long a cache was required. Honestly, I didn't know where to start, and will probably work on doing harder leetcodes in the future to get this sort of thing. The conditions were really crazy.