Day 15
I'm pretty familiar with hashing and buckets as I've learned it in database classes before, so I personally like that I could get this one down pretty easily. Just a couple of math things and list logistics to really hammer out at this point in the tree, but it's really good.