Day 16
It turns out that doing Part 2 as pretty much a copy paste of Part 1 is ok but incredibly slow. It does help though, that it's only 400 permutations of BFS that you need to check. I believe I did make a mistake by changing which direction is "r" and "c", but it ended up not mattering later on after I found the error, because all I really had to do was switch the directions.

Anyway, this wasn't that bad of a problem per se, just time consuming to fill out the BFS portion.