day8
Grid questions are never my strong spot.

Except this really is just me fumbling around looking for the equations when it's really simple in disguise on how to calculate the correct points. I cleaned up the code (a crime) for this question so make it a little easier to parse out what's going on, and I also may have made a logistical optimization by adding it to a queue, which does add memory because yeah, why not.