day20
Took me an extremely long time on part 2 to really figure out that the cheats are direction agnostic and it truly doesn't matter where you are as long as you can cheat. Part 1 was an incredibly proto solution that then required so much overhaul in part 2, please don't pay attention that, it was garbage tier.