day21, part 1
I think this is where I fall off. I was able to get the DP solution at the very least, but it was quite bad in general and I have to find the bottom up solution for part 2. Meanwhile, I'm going to go pack, so this is probably going to be it for me for the year.